Move the \App\Core\ClassicModel to BaseModel and restructure the Demo Module
